% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_counts.R
\name{read_counts}
\alias{read_counts}
\title{Read local census}
\usage{
read_counts(folder, habitat, ecotype)
}
\arguments{
\item{habitat}{What habitat?}

\item{ecotype}{What ecotype?}

\item{filename}{Path to the folder}
}
\description{
Read local census
}
