% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_loci.R
\name{read_loci}
\alias{read_loci}
\title{Read locus-specific data}
\usage{
read_loci(folder, variable)
}
\arguments{
\item{variable}{What to read}

\item{filename}{Path to the folder}
}
\description{
Read locus-specific data
}
