% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_paramfile.R
\name{read_paramfile}
\alias{read_paramfile}
\title{Read parameters}
\usage{
read_paramfile(file, parnames = NULL)
}
\arguments{
\item{file}{Path to the parameter file}

\item{parnames}{Optional vector of parameter names}
}
\description{
Read parameters
}
